테스트

+++ petopia/WebContent/views/info/shareDetailView.jsp
@@ -104,6 +104,7 @@
     	background-color : rgb(255, 248, 240);
 	}
 	
+	/* 댓글 */
 	#comment-table th{text-align : center;}
 	
 	.comment-btn{
@@ -123,14 +124,10 @@
            <div id="edit" align="right">
                 <h4>⋮</h4>
                 <div id="edit-option" style="display:none">
-                <!-- 로그인 안 하면 아무것도 안 보임 / 내 글 : 수정, 삭제 / 다른 사람 글 : 신고 -->
-                <% if(userInfo != null) { %>
-                	<% if(userInfo.getUserNickname().equals(in.getInfoWriter())) {  %>
-                    	<a href="<%= contextPath %>/updateForm.in?ino=<%= in.getInfoNo() %>">수정</a>
-                    	<a href="<%= contextPath %>/delete.in?ictg=<%= in.getCategoryNo() %>&ino=<%= in.getInfoNo() %>">삭제</a>
-                	<% } else {%>
-                    	<a href="#">신고</a>
-                    <% } %>
+                <!-- 로그인 안 하면 아무것도 안 보임 / 내 글 : 수정, 삭제 -->
+                <% if(userInfo != null && userInfo.getUserNickname().equals(in.getInfoWriter())) { %>
+                    <a href="<%= contextPath %>/updateForm.in?ino=<%= in.getInfoNo() %>">수정</a>
+                    <a href="<%= contextPath %>/delete.in?ictg=<%= in.getCategoryNo() %>&ino=<%= in.getInfoNo() %>">삭제</a>
                 <% } %>
                 </div>
            </div>
